fdo#74573: Skip deletion of destination area when 'skip empty' is on.
Also, adjust handling of mix document aka paste functions with this
change.  When using paste function (add, subtract, etc), the behavior
between the 'skip empty' flag on and off makes no difference.  Let's
set the flag to off when paste function is used.

(cherry picked from commit 0c12aa670b83b76241077dfb8bc21f40a55b1667)

+++ sc/source/core/data/document.cxx
@@ -2691,7 +2691,8 @@ void ScDocument::CopyFromClip( const ScRange& rDestRange, const ScMarkData& rMar
         SCCOL nCol2 = pRange->aEnd.Col();
         SCROW nRow2 = pRange->aEnd.Row();
 
-        DeleteArea(nCol1, nRow1, nCol2, nRow2, rMark, nDelFlag);
+        if (!bSkipAttrForEmpty)
+            DeleteArea(nCol1, nRow1, nCol2, nRow2, rMark, nDelFlag);
 
         SCCOL nC1 = nCol1;
         SCROW nR1 = nRow1;
